add create stars function to level scene
